+++ broker/bam/src/ba_ratio_percent.cc
@@ -47,8 +47,8 @@ static constexpr double eps = 0.000001;
  */
 ba_ratio_percent::ba_ratio_percent(
     uint32_t id,
-    uint32_t host_id,
-    uint32_t service_id,
+    uint64_t host_id,
+    uint64_t service_id,
     bool generate_virtual_status,
     const std::shared_ptr<spdlog::logger>& logger)
     : ba(id,
