+++ broker/bam/src/configuration/ba.cc
@@ -91,7 +91,7 @@ uint32_t ba::get_id() const {
  *
  *  @return BA host ID.
  */
-uint32_t ba::get_host_id() const {
+uint64_t ba::get_host_id() const {
   return _host_id;
 }
 
@@ -100,7 +100,7 @@ uint32_t ba::get_host_id() const {
  *
  *  @return  An integer representing the value of this id.
  */
-uint32_t ba::get_service_id() const {
+uint64_t ba::get_service_id() const {
   return _service_id;
 }
 
@@ -163,7 +163,7 @@ ba::downtime_behaviour ba::get_downtime_behaviour() const {
  *
  *  @param[in] service_id  Set the service id.
  */
-void ba::set_host_id(uint32_t host_id) {
+void ba::set_host_id(uint64_t host_id) {
   _host_id = host_id;
 }
 
@@ -172,7 +172,7 @@ void ba::set_host_id(uint32_t host_id) {
  *
  *  @param[in] service_id  Set the service id.
  */
-void ba::set_service_id(uint32_t service_id) {
+void ba::set_service_id(uint64_t service_id) {
   _service_id = service_id;
 }
